refactor: components directory structure
