+++ src/components/index.tsx
@@ -2,26 +2,26 @@
 export {default as App} from './App';
 
 // Typographies
-export {default as H1} from './typography/Heading1';
-export {default as H2} from './typography/Heading2';
-export {default as H3} from './typography/Heading3';
-export {default as H4} from './typography/Heading4';
-export {default as P1} from './typography/Para1';
-export {default as P2} from './typography/Para2';
-export {default as P3} from './typography/Para3';
-export {default as FancyText} from './typography/FancyText';
+export {default as H1} from './atoms/typography/Heading1';
+export {default as H2} from './atoms/typography/Heading2';
+export {default as H3} from './atoms/typography/Heading3';
+export {default as H4} from './atoms/typography/Heading4';
+export {default as P1} from './atoms/typography/Para1';
+export {default as P2} from './atoms/typography/Para2';
+export {default as P3} from './atoms/typography/Para3';
+export {default as FancyText} from './atoms/typography/FancyText';
 
 // Marginals
-export {default as HamburgerMenu} from './marginals/HamburgerMenu';
-export {default as DesktopNavbar} from './marginals/NavbarDesktop';
-export {default as MobileNavbar} from './marginals/NavbarMobile';
-export {default as Navbar} from './marginals/Navbar';
-export {default as Footer} from './marginals/Footer';
+export {default as HamburgerMenu} from './molecules/marginals/HamburgerMenu';
+export {default as DesktopNavbar} from './molecules/marginals/NavbarDesktop';
+export {default as MobileNavbar} from './molecules/marginals/NavbarMobile';
+export {default as Navbar} from './molecules/marginals/Navbar';
+export {default as Footer} from './molecules/marginals/Footer';
 
 // Layout
-export {default as Container} from './layout/Container';
-export {default as PageContainer} from './layout/PageContainer';
-export {default as Layout} from './layout/Layout';
+export {default as Container} from './atoms/layout/Container';
+export {default as PageContainer} from './atoms/layout/PageContainer';
+export {default as Layout} from './atoms/layout/Layout';
 
 // Atoms
 export {default as InAppLink} from './atoms/links/InAppLink';
@@ -37,7 +37,7 @@ export {default as ThemeToggle} from './atoms/ThemeToggle';
 export * from './atoms/ListItems/Lists';
 
 // Work Experience
-export {default as WorkExperience} from './work';
+export {default as WorkExperience} from './molecules/work';
 
 // Projects
-export {default as FeaturedProject} from './projects/FeaturedProject';
+export {default as FeaturedProject} from './molecules/projects/FeaturedProject';
