+++ src/components/molecules/projects/FeaturedProject.tsx
